+++ public_api_client/src/util/messages.js
@@ -1,0 +1,9 @@
+export const MESSAGES = {
+  APP_NAME: 'Fiat Currencies',
+  LOGOUT: 'Logout',
+  LOGIN: 'Login',
+  REGISTER: 'Register',
+  SOMETHING_WRONG: 'Something went wrong!',
+  LOGIN_PAGE: 'Login Page',
+  REGISTER_PAGE: 'Register below to create an account.',
+};
